server/apps/uav/tests/requirements.txt
pytest==7.4.3
pact-python==1.7.0
requests==2.31.0
pytest-asyncio==0.21.1 